+++ include/utils.h
@@ -1,6 +1,8 @@
 #pragma once
 
+#include "types.h"
+
 class Utils{
   public:
-    static int calculateLeds(int noiseMeasurement, int minMeasurement, int maxMeasurement, int steps);
+    static uint16_t calculateLeds(uint16_t noiseMeasurement, uint16_t minMeasurement, uint16_t maxMeasurement, uint16_t steps);
 };
